+++ src/main/java/com/idle/idle_backend/walk/entity/Walk.java
@@ -29,7 +29,7 @@ public class Walk {
 
     private String energyFinishTime;
 
-    private Long energyFinishDistance;
+    private Float energyFinishDistance;
 
     @CreatedDate
     private LocalDateTime createdDate;
